+++ app/components/layout/header/types.ts
@@ -14,16 +14,26 @@ export type HeaderMenuProps = {
   primaryDomainUrl: HeaderProps['header']['shop']['primaryDomain']['url'];
   viewport: Viewport;
   publicStoreDomain: HeaderProps['publicStoreDomain'];
+  cart: Promise<CartApiQueryFragment | null>;
+  shopName: string;
 };
 
 export type MenuDesktopProps = {
   menu: HeaderProps['header']['menu'];
   primaryDomainUrl: HeaderProps['header']['shop']['primaryDomain']['url'];
   publicStoreDomain: HeaderProps['publicStoreDomain'];
+  cart: Promise<CartApiQueryFragment | null>;
+  shopName: string;
+  isLoggedIn: Promise<boolean>;
 };
 
 export type MenuMobileProps = {
   menu: HeaderProps['header']['menu'];
   primaryDomainUrl: HeaderProps['header']['shop']['primaryDomain']['url'];
   publicStoreDomain: HeaderProps['publicStoreDomain'];
+};
+
+export type HeaderActionsProps = {
+  isLoggedIn: Promise<boolean>;
+  cart: Promise<CartApiQueryFragment | null>;
 };
