--- app/components/shared/Icon.tsx
@@ -1,3 +1,0 @@
-export const Icon = ({icon, id}: {id: string; icon: string}) => {
-  return <div key={id}>{icon}</div>;
-};
