+++ app/components/shared/cart/CartBadge.tsx
@@ -1,0 +1,26 @@
+import type {CartViewPayload} from '@shopify/hydrogen';
+import {useAnalytics} from '@shopify/hydrogen';
+import {useAside} from '~/providers/Aside';
+
+export function CartBadge({count}: {count: number | null}) {
+  const {open} = useAside();
+  const {publish, shop, cart, prevCart} = useAnalytics();
+
+  return (
+    <a
+      href="/cart"
+      onClick={(e) => {
+        e.preventDefault();
+        open('cart');
+        publish('cart_viewed', {
+          cart,
+          prevCart,
+          shop,
+          url: window.location.href || '',
+        } as CartViewPayload);
+      }}
+    >
+      Cart {count === null ? <span>&nbsp;</span> : count}
+    </a>
+  );
+}
