+++ app/components/shared/cart/types.ts
@@ -1,0 +1,3 @@
+import type {HeaderProps} from '~/components/layout/header/types';
+
+export type CartToggleProps = Pick<HeaderProps, 'cart'>;
